+++ tests/screenshot/apps/components/ProgressBar.js
@@ -4,8 +4,7 @@ import React from 'react';
 const ProgressBarTests = [
 	<ProgressBar />,
 	<ProgressBar orientation="vertical" />,
-	<ProgressBar orientation="radial" />,
-	<ProgressBar progress={0.75} highlighted />,
+	<ProgressBar progress={0.75} />,
 	<ProgressBar progress={0.75} size="small" />,
 	<ProgressBar progress={0.75} disabled />,
 	<ProgressBar progress={0.75} progressAnchor={0.5} />